refactor: extract shared ACP adapter infrastructure and Hermes UI
Extract duplicated logic from CursorAdapter and HermesAdapter into
AcpAdapterCore: settlement helpers, mode resolution, resume token
parsing, and auto-approval selection. Reduces HermesAdapter from
986 to 861 lines and CursorAdapter from 1123 to 1004 lines.

Extract HermesSetupSection from ProviderInstanceCard to keep the
file under 1000 lines (1000 to 925).

Move Hermes chat surface CSS to dedicated hermes-theme.css for
feature-scoped styling.
