+++ apps/web/src/components/settings/ProviderInstanceCard.tsx
@@ -47,6 +47,7 @@ import {
   getProviderVersionLabel,
   type ProviderStatusKey,
 } from "./providerStatus";
+import { HermesSetupSection } from "./HermesSetupSection";
 
 const PROVIDER_ACCENT_SWATCHES = [
   "#2563eb",
@@ -399,39 +400,6 @@ function ProviderEnvironmentSection(props: {
   );
 }
 
-function ProviderSetupCommandRow(props: {
-  readonly command: string;
-  readonly label: string;
-  readonly onCopy: (command: string, label: string) => void;
-}) {
-  return (
-    <div className="flex min-w-0 items-center gap-1 rounded-md border border-border/70 bg-muted/30 py-0.5 pr-0.5 pl-2">
-      <ScrollArea scrollFade className="h-8 min-w-0 flex-1 rounded-none">
-        <code className="flex h-full w-max items-center whitespace-nowrap pr-3 font-mono text-[11px] text-foreground">
-          {props.command}
-        </code>
-      </ScrollArea>
-      <Tooltip>
-        <TooltipTrigger
-          render={
-            <Button
-              type="button"
-              size="icon-xs"
-              variant="ghost"
-              className="size-6 shrink-0 rounded-sm p-0 text-muted-foreground hover:text-foreground"
-              onClick={() => props.onCopy(props.command, props.label)}
-              aria-label={`Copy ${props.label}`}
-            >
-              <CopyIcon className="size-3" />
-            </Button>
-          }
-        />
-        <TooltipPopup side="top">Copy command</TooltipPopup>
-      </Tooltip>
-    </div>
-  );
-}
-
 interface ProviderInstanceCardProps {
   readonly instanceId: ProviderInstanceId;
   readonly instance: ProviderInstanceConfig;
@@ -735,54 +703,11 @@ export function ProviderInstanceCard({
   ) : null;
 
   const hermesSetupNode = isHermesDriver ? (
-    <div className="border-t border-border/60 px-4 py-3 sm:px-5">
-      <div className="grid gap-3">
-        <div className="grid gap-1">
-          <span className="text-xs font-medium text-foreground">Hermes setup</span>
-          <p className="text-xs leading-snug text-muted-foreground">
-            T3 Code starts Hermes through ACP. Configure Hermes once, then verify the ACP command
-            before sending a turn.
-          </p>
-        </div>
-        {suggestedBinaryPath ? (
-          <div className="grid gap-2 rounded-md border border-border/70 bg-muted/20 p-2">
-            <span className="text-xs text-muted-foreground">
-              Detected Hermes at <code className="text-foreground">{suggestedBinaryPath}</code>
-            </span>
-            <Button
-              type="button"
-              size="xs"
-              variant="outline"
-              className="w-fit"
-              onClick={() => applySuggestedBinaryPath(suggestedBinaryPath)}
-              aria-label="Use detected Hermes path"
-            >
-              Use detected path
-            </Button>
-          </div>
-        ) : null}
-        <div className="grid gap-2">
-          <ProviderSetupCommandRow
-            command="hermes model"
-            label="Hermes setup command"
-            onCopy={(command, label) => copyToClipboard(command, { providerName: label })}
-          />
-          <ProviderSetupCommandRow
-            command="hermes acp"
-            label="Hermes ACP verification command"
-            onCopy={(command, label) => copyToClipboard(command, { providerName: label })}
-          />
-        </div>
-        <a
-          href="docs/providers/hermes.md"
-          target="_blank"
-          rel="noreferrer"
-          className="w-fit text-xs font-medium text-primary hover:underline"
-        >
-          Hermes setup docs
-        </a>
-      </div>
-    </div>
+    <HermesSetupSection
+      suggestedBinaryPath={suggestedBinaryPath}
+      onApplySuggestedPath={applySuggestedBinaryPath}
+      onCopyCommand={(command, label) => copyToClipboard(command, { providerName: label })}
+    />
   ) : null;
 
   return (
